requirements.txt
packaging
torch
metrics
transformers>=4.41.2
datasets>=2.16.0
accelerate>=0.30.1
peft>=0.11.1
trl>=0.8.63
gradio>=4.0.0
numpy
scipy
einops
sentencepiece
tiktoken
protobuf
uvicorn
pydantic
fastapi
sse-starlette
matplotlib>=3.7.0
fire
pyyaml
flash_attn
deepspeed
wandb
alpaca-eval
huggingface-hub
bitsandbytes
evaluate
nvitop
jieba
nltk
rouge-chinese